requirements.txt
aiogram==2.25.1
sqlalchemy==1.4.46
python-dotenv==1.0.0
aiohttp==3.8.5